+++ tests/acp/settings_module_test.php
@@ -9,20 +9,6 @@
 
 namespace blitze\sitemaker\tests\acp;
 
-/**
- * Override trigger_error() in current namespace for testing
- *
- * @return int
- */
-function trigger_error()
-{
-    return false;
-}
-function check_form_key()
-{
-    return true;
-}
-
 use phpbb\request\request_interface;
 use blitze\sitemaker\acp\settings_module;
 
@@ -49,7 +35,6 @@ class settings_module_test extends \phpbb_database_test_case
 		self::$helper = new \phpbb_test_case_helpers(__CLASS__);
 
 		self::$helper->copy_dir($phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/ext/foo/bar', $phpbb_root_path . 'ext/foo/bar');
-		copy($phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/filemanager/test_config.php', $phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/filemanager/config.php');
 	}
 
 	static public function tearDownAfterClass()
@@ -60,6 +45,23 @@ class settings_module_test extends \phpbb_database_test_case
 
 		self::$helper->empty_dir($phpbb_root_path . 'ext/foo');
 		rmdir($phpbb_root_path . 'ext/foo');
+	}
+	
+	protected function setUp()
+	{
+		global $phpbb_root_path;
+
+		parent::setUp();
+
+		copy($phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/filemanager/test_config.php', $phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/filemanager/config.php');
+    }
+
+	protected function tearDown()
+	{
+		global $phpbb_root_path;
+
+		parent::tearDown();
+
 		unlink($phpbb_root_path . 'ext/blitze/sitemaker/tests/acp/fixtures/filemanager/config.php');
 	}
 
@@ -87,11 +89,12 @@ class settings_module_test extends \phpbb_database_test_case
 	 * Get the settings_module object
 	 *
 	 * @param array $variable_map
-	 * @param array $db_text
-	 * @param bool $submitted
+	 * @param array $layout_prefs
+	 * @param array $orphaned_blocks
+	 * @param string $submit_var
 	 * @return \blitze\sitemaker\acp\settings_module
 	 */
-	public function get_module(array $variable_map, array $db_text = array(), $submitted = false)
+	public function get_module(array $variable_map, array $layout_prefs = array(), array $orphaned_blocks = array(), $submit_var = '')
 	{
 		global $phpbb_container, $config, $db, $phpbb_dispatcher, $request, $template, $user, $phpbb_root_path, $phpEx;
 
@@ -99,19 +102,24 @@ class settings_module_test extends \phpbb_database_test_case
 		$phpbb_dispatcher = new \phpbb_mock_event_dispatcher();
 
 		$config = new \phpbb\config\config(array(
-			'sm_hide_birthday'	=> false,
-			'sm_hide_login'		=> false,
-			'sm_hide_online'	=> false,
-			'sm_show_forum_nav'	=> true,
-			'sm_navbar_menu'	=> 2,
-			'sm_forum_icon'		=> 'fa fa-comments',
-			'sm_filemanager'	=> false,
+			'sm_hide_birthday'		=> false,
+			'sm_hide_login'			=> false,
+			'sm_hide_online'		=> false,
+			'sm_show_forum_nav'		=> true,
+			'sm_navbar_menu'		=> 2,
+			'sm_forum_icon'			=> 'fa fa-comments',
+			'sm_filemanager'		=> false,
+			'sitemaker_column_widths' => json_encode([
+				1 => ['sidebar' => '200px'],
+				2 => ['sidebar' => '25%'],
+			]),
+			'sm_orphaned_blocks'	=> json_encode($orphaned_blocks),
 		));
 		$this->config = &$config;
 		$this->php_ext = $phpEx;
 
 		$this->config_text = new \phpbb\config\db_text($db, 'phpbb_config_text');
-		$this->config_text->set('sm_layout_prefs', json_encode($db_text));
+		$this->config_text->set('sm_layout_prefs', json_encode($layout_prefs));
 
 		$request = $this->getMock('\phpbb\request\request_interface');
 		$request->expects($this->any())
@@ -120,7 +128,9 @@ class settings_module_test extends \phpbb_database_test_case
 			->will($this->returnValueMap($variable_map));
 		$request->expects($this->any())
 			->method('is_set_post')
-			->willReturn($submitted);
+			->willReturnCallback(function($var) use ($submit_var) {
+				return $submit_var === $var;
+			});
 
 		$translator = $this->getMockBuilder('\phpbb\language\language')
 			->disableOriginalConstructor()
@@ -132,11 +142,18 @@ class settings_module_test extends \phpbb_database_test_case
 			});
 
 		$user = new \phpbb\user($translator, '\phpbb\datetime');
+		$user->data['user_id'] = 2;
 		$user->data['user_lang'] = 'en';
+		$user->data['user_form_salt'] = '';
 
 		$tpl_data = array();
 		$template = $this->getMockBuilder('\phpbb\template\template')
 			->getMock();
+		$template->expects($this->any())
+			->method('assign_var')
+			->will($this->returnCallback(function($key, $value) use (&$tpl_data) {
+				$tpl_data[$key] = $value;
+			}));
 		$template->expects($this->any())
 			->method('assign_vars')
 			->will($this->returnCallback(function($data) use (&$tpl_data) {
@@ -171,9 +188,22 @@ class settings_module_test extends \phpbb_database_test_case
 			),
 			$phpbb_container);
 
+		$blocks_manager = $this->getMockBuilder('\blitze\sitemaker\services\blocks\manager')
+			->disableOriginalConstructor()
+			->getMock();
+		$blocks_manager->method('get_routes')
+			->willReturn([]);
+		$blocks_manager->method('get_unique_block_names')
+			->willReturn([]);
+
+		$url_checker = $this->getMockBuilder('\blitze\sitemaker\services\url_checker')
+			->getMock();
+
 		$this->icon_picker = $this->getMockBuilder('\blitze\sitemaker\services\icon_picker')
 			->disableOriginalConstructor()
 			->getMock();
+
+		$this->blocks_cleaner = new \blitze\sitemaker\services\blocks\cleaner($this->config, $db, $blocks_manager, $url_checker, '', '');
 
 		$table_prefix = 'phpbb_';
 		$tables = array(
@@ -195,8 +225,16 @@ class settings_module_test extends \phpbb_database_test_case
 		$phpbb_container->set('blitze.sitemaker.icon_picker', $this->icon_picker);
 		$phpbb_container->set('blitze.sitemaker.filemanager.settings', $filemanager);
 		$phpbb_container->set('blitze.sitemaker.mapper.factory', $mapper_factory);
+		$phpbb_container->set('blitze.sitemaker.blocks.cleaner', $this->blocks_cleaner);
 
-		return new \blitze\sitemaker\acp\settings_module(false);
+		$module_service = $this->getMockBuilder('\blitze\sitemaker\acp\settings_module')
+			->setMethods(['check_form_key'])
+			->getMock();
+		$module_service->expects($this->any())
+			->method('check_form_key')
+			->willReturn(true);
+
+		return $module_service;
 	}
 
 	/**
@@ -213,6 +251,10 @@ class settings_module_test extends \phpbb_database_test_case
 						'layout' => 'phpBB/ext/blitze/sitemaker/styles/all/template/layouts/portal/',
 						'view' => 'simple',
 					),
+				),
+				array(
+					'styles' => [2, 5],
+					'routes' => ['https://example.com/phpBB/no_exist.php'],
 				),
 				array(
 					'styles' => array(
@@ -243,6 +285,10 @@ class settings_module_test extends \phpbb_database_test_case
 						'image_max_width'			=> 0,
 						'image_resizing'			=> false,
 					),
+					'orphaned_blocks' => array(
+						'styles' => [2, 5],
+						'routes' => ['https://example.com/phpBB/no_exist.php'],
+					),
 					'sm_user_lang' => 'en',
 					'menu_options' => '<option value="1">Menu 1</option><option value="2" selected="selected">Menu 2</option><option value="3">Menu 3</option>',
 				),
@@ -254,12 +300,13 @@ class settings_module_test extends \phpbb_database_test_case
 	 * Test the main method
 	 *
 	 * @dataProvider module_test_data
-	 * @param array $db_text
+	 * @param array $layout_prefs
+	 * @param array $orphaned_blocks
 	 * @param array $expected
 	 */
-	public function test_module(array $db_text, array $expected)
+	public function test_module(array $layout_prefs, $orphaned_blocks, array $expected)
 	{
-		$module = $this->get_module(array(), $db_text);
+		$module = $this->get_module(array(), $layout_prefs, $orphaned_blocks);
 
 		$this->icon_picker->expects($this->once())
 			->method('picker');
@@ -275,65 +322,141 @@ class settings_module_test extends \phpbb_database_test_case
 	}
 
 	/**
-	 * Test save settings
+	 * Data set for test_save_settings
+	 *
+	 * @return array
 	 */
-	public function test_save_settings()
+	public function save_settings_test_data()
 	{
-		$config = array(
-			'sm_hide_birthday'	=> 1,
-			'sm_hide_login'		=> 1,
-			'sm_hide_online'	=> 1,
-			'sm_show_forum_nav'	=> 1,
-			'sm_navbar_menu'	=> 3,
-			'sm_filemanager'	=> 1,
-			'sm_forum_icon'		=> 'fa fa-car',
-		);
-		$filemanager = array(
-			'image_watermark_coordinates' => '40x50',
-			'image_max_width' => 800,
-			'image_resizing' => 'true',
-			'image_watermark' => './test.png',
-		);
-		$layouts = array(
-			1 => array (
-				'layout' => './../ext/blitze/sitemaker/styles/all/template/layouts/blog/',
-				'view' => 'boxed',
+		return array(
+			array(
+				'orphans',
+				array(
+					array('cleanup', array(0 => ''), false, request_interface::REQUEST, array('styles', 'blocks'))
+				),
+				array(
+					'forum_icon'			=> 'fa fa-comments',
+					'navbar_menu'			=> 2,
+					'show_forum_nav'		=> true,
+					'hide_login'			=> false,
+					'hide_online'			=> false,
+					'hide_birthday'			=> false,
+					'filemanager'			=> false,
+					'filemanager_config'	=> array(
+						'image_watermark'			=> './image.jpg',
+						'image_watermark_position'	=> 'br',
+						'image_max_width'			=> 0,
+						'image_resizing'			=> false,
+					),
+					'layout_prefs'			=> array(
+						1 => array(
+							'layout'	=> 'phpBB/ext/blitze/sitemaker/styles/all/template/layouts/portal/',
+							'view'		=> 'simple',
+						),
+					),
+					'orphaned_blocks'		=> null,
+				),
+				'BLOCKS_CLEANUP_DONE',
 			),
-			2 => array (
-				'layout' => './../ext/foo/bar/styles/prosilver/template/layouts/my_layout/',
+			array(
+				'submit',
+				array(
+					array('config', array('' => ''), false, request_interface::REQUEST, array(
+						'sm_hide_birthday'	=> 1,
+						'sm_hide_login'		=> 1,
+						'sm_hide_online'	=> 1,
+						'sm_show_forum_nav'	=> 1,
+						'sm_navbar_menu'	=> 3,
+						'sm_filemanager'	=> 1,
+						'sm_forum_icon'		=> 'fa fa-car',
+					)),
+					array('filemanager', array('' => ''), false, request_interface::REQUEST, array(
+						'image_watermark_coordinates' => '40x50',
+						'image_max_width' => 800,
+						'image_resizing' => 'true',
+						'image_watermark' => './test.png',
+					)),
+					array('layouts', array(0 => array('' => '')), false, request_interface::REQUEST, array(
+						1 => array (
+							'layout' => 'phpBB/ext/blitze/sitemaker/styles/all/template/layouts/blog/',
+							'view' => 'boxed',
+						),
+						2 => array (
+							'layout' => 'phpBB/ext/foo/bar/styles/prosilver/template/layouts/my_layout/',
+							'view' => 'simple',
+						),
+					)),
+				),
+				array(
+					'forum_icon'			=> 'fa fa-car',
+					'navbar_menu'			=> 3,
+					'show_forum_nav'		=> 1,
+					'hide_login'			=> 1,
+					'hide_online'			=> 1,
+					'hide_birthday'			=> 1,
+					'filemanager'			=> 1,
+					'filemanager_config'	=> array(
+						'image_watermark'			=> './test.png',
+						'image_watermark_position'	=> '40x50',
+						'image_max_width'			=> 800,
+						'image_resizing'			=> true,
+					),
+					'layout_prefs'			=> array(
+						1 => array(
+							'layout'	=> 'phpBB/ext/blitze/sitemaker/styles/all/template/layouts/blog/',
+							'view'		=> 'boxed',
+						),
+						2 => array(
+							'layout'	=> 'phpBB/ext/foo/bar/styles/prosilver/template/layouts/my_layout/',
+							'view'		=> 'simple',
+						),
+					),
+					'orphaned_blocks'		=> array(
+						'styles' => [2, 5],
+						'routes' => ['https://example.com/phpBB/no_exist.php'],
+						'blocks' => ['my.invalid.block'],
+					),
+				),
+				'SETTINGS_SAVED',
+			),
+		);
+	}
+
+	/**
+	 * Test save settings
+	 *
+	 * @dataProvider save_settings_test_data
+	 * @param string $submit_var
+	 * @param array $variable_map
+	 * @param array $expected
+	 * @param string $message
+	 */
+	public function test_save_settings($submit_var, array $variable_map, array $expected, $message)
+	{
+		$layout_prefs = array(
+			1 => array(
+				'layout' => 'phpBB/ext/blitze/sitemaker/styles/all/template/layouts/portal/',
 				'view' => 'simple',
 			),
 		);
-		$variable_map = array(
-			array('config', array('' => ''), false, request_interface::REQUEST, $config),
-			array('filemanager', array('' => ''), false, request_interface::REQUEST, $filemanager),
-			array('layouts', array(0 => array('' => '')), false, request_interface::REQUEST, $layouts),
+
+		$orphaned_blocks = array(
+			'styles' => [2, 5],
+			'routes' => ['https://example.com/phpBB/no_exist.php'],
+			'blocks' => ['my.invalid.block'],
 		);
 
-		$module = $this->get_module($variable_map, array(), true);
-		$reflection = new \ReflectionClass($module);
-		$method = $reflection->getMethod('handle_submit');
-		$method->setAccessible(true);
+		$module = $this->get_module($variable_map, $layout_prefs, $orphaned_blocks, $submit_var);
 
-		$parameters = array('form_key');
-		$method->invokeArgs($module, $parameters);
-
-		$expected = array(
-            'forum_icon'			=> 'fa fa-car',
-			'navbar_menu'			=> 3,
-            'show_forum_nav'		=> 1,
-			'hide_login'			=> 1,
-			'hide_online'			=> 1,
-			'hide_birthday'			=> 1,
-			'layout_prefs'			=> $layouts,
-			'filemanager'			=> 1,
-			'filemanager_config'	=> array(
-				'image_watermark'			=> './test.png',
-				'image_watermark_position'	=> '40x50',
-				'image_max_width'			=> 800,
-				'image_resizing'			=> true,
-			),
-		);
+		try
+		{
+			$module->main();
+		}
+		catch (\Exception $e)
+		{
+			preg_match('/\w+/i', $e->getMessage(), $matches);
+			$this->assertEquals(!empty($matches[0]) ? $matches[0] : '', $message);
+		}
 
 		$result = array(
             'forum_icon'			=> $this->config['sm_forum_icon'],
@@ -342,9 +465,10 @@ class settings_module_test extends \phpbb_database_test_case
 			'hide_login'			=> $this->config['sm_hide_login'],
 			'hide_online'			=> $this->config['sm_hide_online'],
 			'hide_birthday'			=> $this->config['sm_hide_birthday'],
-			'layout_prefs'			=> json_decode($this->config_text->get('sm_layout_prefs'), true),
 			'filemanager'			=> $this->config['sm_filemanager'],
 			'filemanager_config'	=> include($this->config_path . 'config.' . $this->php_ext),
+			'layout_prefs'			=> json_decode($this->config_text->get('sm_layout_prefs'), true),
+			'orphaned_blocks'		=> json_decode($this->config['sm_orphaned_blocks'], true),
 		);
 
 		$this->assertEquals($expected, $result);